add 4/12

+++ week16/example_day2/articles/views.py
@@ -28,7 +28,9 @@ def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -41,8 +43,8 @@ def create(request):
 @login_required
 def delete(request, artilce_pk):
     article = Article.objects.get(pk=artilce_pk)
-
-    article.delete()
+    if request.user == article.user:
+        article.delete()
 
     return redirect('articles:index')
 
@@ -50,13 +52,16 @@ def delete(request, artilce_pk):
 @login_required
 def update(request, article_pk):
     article = Article.objects.get(pk=article_pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:index')
     context = {
         'article': article,
         'form': form,
@@ -69,6 +74,7 @@ def comment_create(request, article_pk):
     if comment_form.is_valid():
         comment = comment_form.save(commit=False)
         comment.article = article
+        comment.user = request.user
         comment.save()
         return redirect('articles:detail', article.pk)
     context = {
@@ -79,5 +85,6 @@ def comment_create(request, article_pk):
 
 def comment_delete(request, article_pk, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
+    if request.user == comment.user:
+        comment.delete()
     return redirect('articles:detail', article_pk)
